--- .github/workflows/main.yml
@@ -1,30 +1,0 @@
-name: Java CI with Gradle
-
-on:
-  push:
-    branches: [ 1.16.5 ]
-  pull_request:
-    branches: [ 1.16.5 ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up JDK 16
-      uses: actions/setup-java@v2
-      with:
-        java-version: '16'
-        distribution: 'adopt'
-        cache: gradle
-    - name: Grant execute permission for gradlew
-      run: chmod +x gradlew
-    - name: Build with Gradle
-      run: ./gradlew build
-    - name: "Upload Artifact"
-      uses: actions/upload-artifact@v1
-      with:
-        name: Builds
-        path: build/libs
