+++ .github/workflows/publish.yml
@@ -1,0 +1,47 @@
+name: Publish
+
+on:
+  push:
+    tags:
+      - '*'
+
+jobs:
+  publish:
+    runs-on: ubuntu-20.04
+    steps:
+
+      - name: Checkout Repository
+        uses: actions/checkout@v2
+
+      - name: Validate Gradle Wrapper
+        uses: gradle/wrapper-validation-action@v1
+        
+      - uses: actions/setup-java@v2
+        with:
+          distribution: 'adopt'
+          java-version: '8'
+
+      - name: Build
+        run: ./gradlew build
+        
+      - name: Publish to Modrinth
+        run: ./gradlew publishModrinth
+        env:
+          MODRINTH: ${{ secrets.MODRINTH_TOKEN }}
+        
+      - name: Publish to Curseforge
+        run: ./gradlew curseforge
+        env:
+          CURSEFORGE: ${{ secrets.CURSEFORGE_TOKEN }}
+        
+      - name: Publish to Infernal Studios Maven
+        run: ./gradlew publish
+        env:
+          MAVEN_USER: ${{ secrete.MAVEN_USER }}
+          MAVEN_TOKEN: ${{ secrets.MAVEN_TOKEN }}
+
+      - name: Create GitHub Release
+        uses: ncipollo/release-action@v1
+        with:
+          artifacts: "./build/libs/*"
+          token: ${{ secrets.GITHUB_TOKEN }}
