+++ README.md
@@ -1,3 +1,2 @@
-According to the Apache License 2.0, I need to lit out my modification to the code
-
-So here it is:
+# Shield Expansion
+ Expanding upon Shields in Minecraft!
