+++ build.properties
@@ -1,0 +1,29 @@
+VERSION=1.1.5
+
+MINECRAFT_VERSION=1.16.5
+
+FORGE_VERSION=36.2.39
+
+MAPPINGS_CHANNEL=official
+MAPPINGS_VERSION=1.16.5
+
+GROUP=org.infernalstudios
+ARTIFACT=shieldexp
+MOD_ID=shieldexp
+ARCHIVES_BASE_NAME=shieldexp
+TITLE=shieldexp
+AUTHOR=Infernal Studios
+
+## Publishing information
+## Modify this before publishing!!
+
+NAME=Shield Expansion
+URL=https://infernalstudios.org/shieldexp
+LICENSE_NAME=Apache License 2.0
+LICENSE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
+SCM_CONNECTION=scm:git:git://github.com/infernalexp/Shield-Expansion.git
+SCM_URL=https://github.com/infernalexp/Shield-Expansion.git
+MAVEN_URL=https://maven.infernalstudios.org/
+
+#MODRINTH_ID=NotARealId
+#CURSEFORGE_ID=696969
