+++ src/main/java/org/infernalstudios/shieldexp/events/FovEvent.java
@@ -3,19 +3,19 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.teampotato.shieldexpansion.events;
+package org.infernalstudios.shieldexp.events;
 
-import com.teampotato.shieldexpansion.ShieldExpansion;
-import com.teampotato.shieldexpansion.access.LivingEntityAccess;
+import org.infernalstudios.shieldexp.ShieldExpansion;
+import org.infernalstudios.shieldexp.access.LivingEntityAccess;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.FOVUpdateEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
