+++ src/main/java/org/infernalstudios/shieldexp/events/ShieldExpansionEvents.java
@@ -3,20 +3,20 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.teampotato.shieldexpansion.events;
+package org.infernalstudios.shieldexp.events;
 
-import com.teampotato.shieldexpansion.ShieldExpansion;
-import com.teampotato.shieldexpansion.access.LivingEntityAccess;
-import com.teampotato.shieldexpansion.init.Config;
+import org.infernalstudios.shieldexp.ShieldExpansion;
+import org.infernalstudios.shieldexp.access.LivingEntityAccess;
+import org.infernalstudios.shieldexp.init.Config;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.attributes.AttributeModifier;
@@ -42,7 +42,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 
 import java.util.Objects;
 
-import static com.teampotato.shieldexpansion.init.ShieldDataLoader.SHIELD_STATS;
+import static org.infernalstudios.shieldexp.init.ShieldDataLoader.SHIELD_STATS;
 
 
 @Mod.EventBusSubscriber(modid = ShieldExpansion.ID, bus = Mod.EventBusSubscriber.Bus.MOD)
@@ -244,7 +244,7 @@ public class ShieldExpansionEvents {
     //reads a shield attribute from the given shield's stats map, or the default map if no map is found
     public static Double getShieldValue(Item item, String value) {
         String key = Objects.requireNonNull(ForgeRegistries.ITEMS.getKey(item)).toString();
-        return SHIELD_STATS.containsKey(key) ? SHIELD_STATS.get(key).get(value) : SHIELD_STATS.get("shieldexpansion:default").get(value);
+        return SHIELD_STATS.containsKey(key) ? SHIELD_STATS.get(key).get(value) : SHIELD_STATS.get("shieldexp:default").get(value);
     }
 
     //increases the current used stamina count of the given player, and removes the blocking state if the given shield's stamina is used up
