+++ src/main/java/org/infernalstudios/shieldexp/events/TooltipEvents.java
@@ -3,19 +3,19 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.teampotato.shieldexpansion.events;
+package org.infernalstudios.shieldexp.events;
 
-import com.teampotato.shieldexpansion.ShieldExpansion;
-import com.teampotato.shieldexpansion.init.Config;
+import org.infernalstudios.shieldexp.ShieldExpansion;
+import org.infernalstudios.shieldexp.init.Config;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
@@ -36,7 +36,7 @@ public class TooltipEvents {
         PlayerEntity player = event.getPlayer();
         if (player != null && Config.isShield(item)) {
             event.getToolTip().add(ITextComponent.nullToEmpty((" ")));
-            event.getToolTip().add(new TranslationTextComponent("shieldexpansion.tooltip.attribute").withStyle(TextFormatting.GRAY));
+            event.getToolTip().add(new TranslationTextComponent("shieldexp.tooltip.attribute").withStyle(TextFormatting.GRAY));
             tooltip(event, item, "cooldownTicks");
             tooltip(event, item, "stamina");
             tooltip(event, item, "speedFactor");
@@ -44,8 +44,8 @@ public class TooltipEvents {
             tooltip(event, item, "parryTicks");
             tooltip(event, item, "blastResistance");
             event.getToolTip().add(ITextComponent.nullToEmpty(" "));
-            if (Screen.hasShiftDown()) event.getToolTip().add(new TranslationTextComponent("shieldexpansion.tooltip.instructions.parry").withStyle(TextFormatting.YELLOW));
-            else event.getToolTip().add(new TranslationTextComponent("shieldexpansion.tooltip.instructions").withStyle(TextFormatting.GRAY));
+            if (Screen.hasShiftDown()) event.getToolTip().add(new TranslationTextComponent("shieldexp.tooltip.instructions.parry").withStyle(TextFormatting.YELLOW));
+            else event.getToolTip().add(new TranslationTextComponent("shieldexp.tooltip.instructions").withStyle(TextFormatting.GRAY));
         }
     }
 
@@ -53,7 +53,7 @@ public class TooltipEvents {
     private void tooltip(ItemTooltipEvent event, Item item, String attribute) {
         if (validate(item, attribute)) {
             Double value = ShieldExpansionEvents.getShieldValue(item, attribute);
-            String fullKey = "shieldexpansion.tooltip.attribute." + attribute.replaceAll("(?<!^)([A-Z])", "_$1").toLowerCase().trim();
+            String fullKey = "shieldexp.tooltip.attribute." + attribute.replaceAll("(?<!^)([A-Z])", "_$1").toLowerCase().trim();
 
             if (attribute.equals("speedFactor"))
                 event.getToolTip().add(new TranslationTextComponent(value < 0.6 ? fullKey + "_slow" : fullKey + "_fast").withStyle(TextFormatting.DARK_GREEN));
