+++ src/main/java/org/infernalstudios/shieldexp/mixin/LivingEntityMixin.java
@@ -3,16 +3,16 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.teampotato.shieldexpansion.mixin;
+package org.infernalstudios.shieldexp.mixin;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
@@ -38,7 +38,7 @@ public abstract class LivingEntityMixin extends Entity {
     }
 
     @Inject(method = "isBlocking", at = @At(value = "RETURN", ordinal = 1), cancellable = true)
-    private void shieldexpansion$isBlocking(CallbackInfoReturnable<Boolean> ci) {
+    private void shieldexp$isBlocking(CallbackInfoReturnable<Boolean> ci) {
         ci.setReturnValue(this.useItem.getItem().getUseDuration(this.useItem) - this.useItemRemaining >= 0);
     }
 }
