+++ src/main/resources/assets/shieldexpansion/lang/zh_cn.json
@@ -1,29 +1,29 @@
 {
-  "shieldexpansion.tooltip.instructions": "按住Shift查看说明",
-  "shieldexpansion.tooltip.instructions.parry": "在受到攻击前一刻举起盾牌以招架攻击！",
+  "shieldexp.tooltip.instructions": "按住Shift查看说明",
+  "shieldexp.tooltip.instructions.parry": "在受到攻击前一刻举起盾牌以招架攻击！",
 
-  "shieldexpansion.tooltip.attribute": "用作盾牌时：",
-  "shieldexpansion.tooltip.attribute.cooldown_ticks": " %ss 防御冷却",
-  "shieldexpansion.tooltip.attribute.speed_factor_slow": " 举盾移速：慢",
-  "shieldexpansion.tooltip.attribute.speed_factor_fast": " 举盾移速：快",
-  "shieldexpansion.tooltip.attribute.parry_damage": " %s 招架反伤",
-  "shieldexpansion.tooltip.attribute.parry_ticks": " %ss 招架判定窗口",
-  "shieldexpansion.tooltip.attribute.stamina": " %s 防御耐力",
-  "shieldexpansion.tooltip.attribute.blast_resistance": " %s 爆炸抗性",
+  "shieldexp.tooltip.attribute": "用作盾牌时：",
+  "shieldexp.tooltip.attribute.cooldown_ticks": " %ss 防御冷却",
+  "shieldexp.tooltip.attribute.speed_factor_slow": " 举盾移速：慢",
+  "shieldexp.tooltip.attribute.speed_factor_fast": " 举盾移速：快",
+  "shieldexp.tooltip.attribute.parry_damage": " %s 招架反伤",
+  "shieldexp.tooltip.attribute.parry_ticks": " %ss 招架判定窗口",
+  "shieldexp.tooltip.attribute.stamina": " %s 防御耐力",
+  "shieldexp.tooltip.attribute.blast_resistance": " %s 爆炸抗性",
 
-  "item.shieldexpansion.wooden_shield": "木盾牌",
-  "item.shieldexpansion.golden_shield": "金盾牌",
-  "item.shieldexpansion.iron_shield": "铁盾牌",
-  "item.shieldexpansion.diamond_shield": "钻石盾牌",
-  "item.shieldexpansion.netherite_shield": "下界合金盾牌",
-  "item.shieldexpansion.paragon_shield": "圣尊盾牌",
-  "item.shieldexpansion.griefer_shield": "破坏者盾牌",
+  "item.shieldexp.wooden_shield": "木盾牌",
+  "item.shieldexp.golden_shield": "金盾牌",
+  "item.shieldexp.iron_shield": "铁盾牌",
+  "item.shieldexp.diamond_shield": "钻石盾牌",
+  "item.shieldexp.netherite_shield": "下界合金盾牌",
+  "item.shieldexp.paragon_shield": "圣尊盾牌",
+  "item.shieldexp.griefer_shield": "破坏者盾牌",
 
   "death.attack.partialblast": "%1$s的盾牌没能在爆炸中保护好主人",
 
-  "shieldexpansion.advancements.get_diamond_shield.title": "不灭钻石",
-  "shieldexpansion.advancements.get_diamond_shield.description": "合成一面钻石盾牌",
+  "shieldexp.advancements.get_diamond_shield.title": "不灭钻石",
+  "shieldexp.advancements.get_diamond_shield.description": "合成一面钻石盾牌",
 
-  "shieldexpansion.advancements.get_netherite_shield.title": "无法撼动的物体",
-  "shieldexpansion.advancements.get_netherite_shield.description": "合成一面下界合金盾牌"
+  "shieldexp.advancements.get_netherite_shield.title": "无法撼动的物体",
+  "shieldexp.advancements.get_netherite_shield.description": "合成一面下界合金盾牌"
 }
