+++ src/main/resources/assets/shieldexpansion/models/item/diamond_shield.json
@@ -1,8 +1,8 @@
 {
 	"credit": "Nekomaster1000",
 	"textures": {
-		"1": "shieldexpansion:diamond",
-		"particle": "shieldexpansion:diamond"
+		"1": "shieldexp:diamond",
+		"particle": "shieldexp:diamond"
 	},
 	"elements": [
 		{
@@ -31,7 +31,7 @@
 		}
 	],
 	"overrides": [
-		{"predicate": {"blocking": 1}, "model": "shieldexpansion:item/diamond_shield_blocking"}
+		{"predicate": {"blocking": 1}, "model": "shieldexp:item/diamond_shield_blocking"}
 	],
 	"display": {
 		"thirdperson_righthand": {
