+++ src/main/resources/assets/shieldexpansion/models/item/diamond_shield_blocking.json
@@ -1,8 +1,8 @@
 {
 	"credit": "Nekomaster1000",
 	"textures": {
-		"1": "shieldexpansion:diamond",
-		"particle": "shieldexpansion:diamond"
+		"1": "shieldexp:diamond",
+		"particle": "shieldexp:diamond"
 	},
 	"elements": [
 		{
