+++ src/main/resources/assets/shieldexpansion/models/item/golden_shield_blocking.json
@@ -1,8 +1,8 @@
 {
 	"credit": "Nekomaster1000",
 	"textures": {
-		"1": "shieldexpansion:golden",
-		"particle": "shieldexpansion:golden"
+		"1": "shieldexp:golden",
+		"particle": "shieldexp:golden"
 	},
 	"elements": [
 		{
