+++ src/main/resources/assets/shieldexpansion/models/item/griefer_shield_blocking.json
@@ -1,8 +1,8 @@
 {
 	"credit": "Nekomaster1000",
 	"textures": {
-		"1": "shieldexpansion:griefer",
-		"particle": "shieldexpansion:griefer"
+		"1": "shieldexp:griefer",
+		"particle": "shieldexp:griefer"
 	},
 	"elements": [
 		{
