+++ src/main/resources/assets/shieldexpansion/models/item/iron_shield_blocking.json
@@ -1,8 +1,8 @@
 {
 	"credit": "Nekomaster1000",
 	"textures": {
-		"1": "shieldexpansion:iron",
-		"particle": "shieldexpansion:iron"
+		"1": "shieldexp:iron",
+		"particle": "shieldexp:iron"
 	},
 	"elements": [
 		{
