+++ src/main/resources/assets/shieldexpansion/models/item/paragon_shield.json
@@ -1,8 +1,8 @@
 {
 	"credit": "Nekomaster1000",
 	"textures": {
-		"2": "shieldexpansion:paragon",
-		"particle": "shieldexpansion:paragon"
+		"2": "shieldexp:paragon",
+		"particle": "shieldexp:paragon"
 	},
 	"elements": [
 		{
@@ -31,7 +31,7 @@
 		}
 	],
 	"overrides": [
-		{"predicate": {"blocking": 1}, "model": "shieldexpansion:item/paragon_shield_blocking"}
+		{"predicate": {"blocking": 1}, "model": "shieldexp:item/paragon_shield_blocking"}
 	],
 	"display": {
 		"thirdperson_righthand": {
