+++ src/main/resources/assets/shieldexpansion/models/item/wooden_shield.json
@@ -1,8 +1,8 @@
 {
 	"credit": "Nekomaster1000",
 	"textures": {
-		"1": "shieldexpansion:wooden",
-		"particle": "shieldexpansion:wooden"
+		"1": "shieldexp:wooden",
+		"particle": "shieldexp:wooden"
 	},
 	"elements": [
 		{
@@ -35,7 +35,7 @@
 			"predicate": {
 				"blocking": 1
 			},
-			"model": "shieldexpansion:item/wooden_shield_blocking"
+			"model": "shieldexp:item/wooden_shield_blocking"
 		}
 	],
 	"display": {
