+++ src/main/resources/data/shieldexpansion/advancements/get_netherite_shield.json
@@ -1,12 +1,12 @@
 {
-  "parent": "shieldexpansion:get_diamond_shield",
+  "parent": "shieldexp:get_diamond_shield",
   "criteria": {
     "netherite_shield": {
       "conditions": {
         "items": [
           {
             "items": [
-              "shieldexpansion:netherite_shield"
+              "shieldexp:netherite_shield"
             ]
           }
         ]
@@ -20,16 +20,16 @@
   "display": {
     "announce_to_chat": true,
     "description": {
-      "translate": "shieldexpansion.advancements.get_netherite_shield.description"
+      "translate": "shieldexp.advancements.get_netherite_shield.description"
     },
     "frame": "task",
     "hidden": false,
     "icon": {
-      "item": "shieldexpansion:netherite_shield"
+      "item": "shieldexp:netherite_shield"
     },
     "show_toast": true,
     "title": {
-      "translate": "shieldexpansion.advancements.get_netherite_shield.title"
+      "translate": "shieldexp.advancements.get_netherite_shield.title"
     }
   },
   "requirements": [
