+++ src/main/resources/data/shieldexpansion/advancements/recipes/diamond_shield.json
@@ -15,7 +15,7 @@
     },
     "has_the_recipe": {
       "conditions": {
-        "recipe": "shieldexpansion:diamond_shield"
+        "recipe": "shieldexp:diamond_shield"
       },
       "trigger": "minecraft:recipe_unlocked"
     }
@@ -28,7 +28,7 @@
   ],
   "rewards": {
     "recipes": [
-      "shieldexpansion:diamond_shield"
+      "shieldexp:diamond_shield"
     ]
   }
 }
