+++ src/main/resources/data/shieldexpansion/advancements/recipes/golden_shield.json
@@ -15,7 +15,7 @@
     },
     "has_the_recipe": {
       "conditions": {
-        "recipe": "shieldexpansion:golden_shield"
+        "recipe": "shieldexp:golden_shield"
       },
       "trigger": "minecraft:recipe_unlocked"
     }
@@ -28,7 +28,7 @@
   ],
   "rewards": {
     "recipes": [
-      "shieldexpansion:golden_shield"
+      "shieldexp:golden_shield"
     ]
   }
 }
