+++ src/main/resources/data/shieldexpansion/recipes/diamond_shield.json
@@ -14,6 +14,6 @@
       }
     },
     "result": {
-      "item": "shieldexpansion:diamond_shield"
+      "item": "shieldexp:diamond_shield"
     }
   }
