+++ src/main/resources/data/shieldexpansion/recipes/golden_shield.json
@@ -14,6 +14,6 @@
       }
     },
     "result": {
-      "item": "shieldexpansion:golden_shield"
+      "item": "shieldexp:golden_shield"
     }
   }
