+++ src/main/resources/data/shieldexpansion/recipes/griefer_shield.json
@@ -14,6 +14,6 @@
     }
   },
   "result": {
-    "item": "shieldexpansion:griefer_shield"
+    "item": "shieldexp:griefer_shield"
   }
 }
