+++ src/main/resources/data/shieldexpansion/recipes/iron_shield.json
@@ -14,6 +14,6 @@
       }
     },
     "result": {
-      "item": "shieldexpansion:iron_shield"
+      "item": "shieldexp:iron_shield"
     }
   }
