+++ src/main/resources/data/shieldexpansion/recipes/wooden_shield.json
@@ -14,6 +14,6 @@
       }
     },
     "result": {
-      "item": "shieldexpansion:wooden_shield"
+      "item": "shieldexp:wooden_shield"
     }
   }
